openai_assistant/pyproject.toml
[tool.poetry]
name = "chainlit-template"
version = "0.4.3"
description = ""
authors = ["BCM <dahifi@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
chainlit = "^1.0.505"
openai = "^1.23.3"
mailchimp-marketing = "^3.0.80"
pyyaml = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


